# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Goatup Chat API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group info (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/members:
    post:
      operationId: addGroupMember
      tags: [groups]
      summary: Add a member to the group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMemberInput"
      responses:
        "200":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/members/{userId}:
    delete:
      operationId: removeGroupMember
      tags: [groups]
      summary: Remove a member from the group (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/members/{userId}/promote:
    patch:
      operationId: promoteGroupMember
      tags: [groups]
      summary: Promote member to admin
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Promoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /groups/{groupId}/leave:
    post:
      operationId: leaveGroup
      tags: [groups]
      summary: Leave a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Blocks
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
...
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: ["string", "null"]
        memberIds:
          type: array
          items:
            type: integer
    GroupUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
    GroupMemberInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
    GroupDetail:
      type: object
      required: [id, name, conversationId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        conversationId:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMember"
        createdAt:
          type: string
    GroupMember:
      type: object
      required: [userId, username, isAdmin, joinedAt]
      properties:
        userId:
          type: integer
        username:
          type: string
        profilePicUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        joinedAt:
          type: string

    # Block schemas
...