# lib/api-zod/src/generated/types/groupDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Goatup Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { GroupMember } from './groupMember';

export interface GroupDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  conversationId: number;
  members?: GroupMember[];
  createdAt: string;
}