# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Goatup Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /** @nullable */
  description?: string | null;
  memberIds?: number[];
}