# lib/api-zod/src/generated/types/groupUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Goatup Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconUrl?: string | null;
}